created catalog page
