+++ src/entities/product-card/ui/media/index.ts
