--- src/pages/catalog-page/catalog-page.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function CatalogPage() {
-  return <div>CatalogPage</div>;
-}
