+++ src/pages/catalog-page/ui/catalog-card/card-tabs/index.ts
@@ -1,0 +1,1 @@
+export { CatalogCardTabs } from "./catalog-card-tabs";
