+++ src/pages/catalog-page/ui/catalog-card/delivery/delivery.tsx
