+++ src/pages/catalog-page/ui/catalog-card/delivery/index.ts
