+++ src/pages/catalog-page/ui/catalog-card/features/add-to-favorites-button.tsx
