+++ src/pages/catalog-page/ui/catalog-card/index.ts
@@ -1,0 +1,1 @@
+export { CatalogCard } from "./catalog-card";
