+++ src/pages/catalog-page/ui/catalog-card/sale.tsx
