+++ src/pages/catalog-page/ui/catalog-card/seller/index.ts
