+++ src/pages/catalog-page/ui/catalog-card/types/slot.ts
