+++ src/pages/catalog-page/ui/catalog-list/index.ts
@@ -1,0 +1,1 @@
+export { CatalogList } from "./catalog-list";
