+++ src/pages/catalog-page/ui/catalog-list/products.json
