+++ src/pages/catalog-page/ui/catalog-list/types/delivery.ts
