+++ src/pages/catalog-page/ui/catalog-list/types/index.ts
