+++ src/pages/catalog-page/ui/catalog-list/types/product-value.ts
