+++ src/pages/catalog-page/ui/catalog-list/types/sale.ts
