+++ src/pages/catalog-page/ui/catalog-list/types/seller.ts
